Check latest stable version on winsparkle.
Due to how we generate build version for nightlies, we have something
like `N.N.N-"commit hash"`. When using winsparkle, the nightly believes
there is an update, for version `N.N.N`. This is not how it works, so
we remove the dash and everything after for our update check on
winsparkle.

It is not clear if this is a bug on winsparkle or if we are abusing
our definition of version to generate a build from a stable
version with a bunch of patches applied to it (our WIP for next).